chat_api/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0  # Добавьте [standard] для дополнительных зависимостей
httpx>=0.23.0
python-dotenv>=0.19.0
aiohttp
